requirements.txt
fastapi==0.110.0
uvicorn[standard]
python-dotenv
oauth2client
gspread